Implement ForecastDataAverageByDayAggregator

+++ src/WeatherApp.BusinessLayer/Shared/ForecastData.cs
@@ -5,7 +5,7 @@ namespace WeatherApp.BusinessLayer.Shared
     public class ForecastData
     {
         public double Temperature { get; set; }
-        public int Humidity { get; set; }
+        public double Humidity { get; set; }
         public double WindSpeed { get; set; }
         public DateTime TimeStamp { get; set; }
     }
